backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
google-cloud-storage==2.12.0
shapely==2.0.6
pydantic==2.5.2
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
numpy<2.0.0
# Note: bson is included with pymongo, separate installation not needed